requirements.txt
fastapi
uvicorn
cloudscraper
beautifulsoup4
yagmail
psycopg2-binary
requests